+++ spec/requests/relationships_spec.rb
@@ -109,50 +109,4 @@ RSpec.describe RelationshipsController, type: :request do
       end
     end
   end
-
-  describe 'GET /users/:user_id/followees' do
-    let!(:relationship) { create(:relationship, follower: user, followee: other_user) }
-
-    context 'ログイン状態のとき' do
-      include_context 'userでログインする'
-
-      it 'followees一覧が取得できる' do
-        get user_followees_path(user)
-        expect(response.body).to include(other_user.email)
-      end
-    end
-
-    context 'ログイン状態でないとき' do
-      it 'ログインページにリダイレクトされる' do
-        get user_followees_path(user)
-        aggregate_failures do
-          expect(flash[:alert]).to eq 'ログインしてください'
-          expect(response).to redirect_to(login_url)
-        end
-      end
-    end
-  end
-
-  describe 'GET /users/:user_id/followers' do
-    let!(:relationship) { create(:relationship, follower: other_user, followee: user) }
-
-    context 'ログイン状態のとき' do
-      include_context 'userでログインする'
-
-      it 'followers一覧が取得できる' do
-        get user_followers_path(user)
-        expect(response.body).to include(other_user.email)
-      end
-    end
-
-    context 'ログイン状態でないとき' do
-      it 'ログインページにリダイレクトされる' do
-        get user_followees_path(user)
-        aggregate_failures do
-          expect(flash[:alert]).to eq 'ログインしてください'
-          expect(response).to redirect_to(login_url)
-        end
-      end
-    end
-  end
 end
